+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/di/AppComponent.kt
@@ -1,7 +1,6 @@
 package com.github.klee0kai.thekey.app.di
 
 import android.app.Activity
-import androidx.activity.OnBackPressedDispatcher
 import androidx.compose.material3.SnackbarHostState
 import com.github.klee0kai.stone.KotlinWrappersStone
 import com.github.klee0kai.stone.Stone
@@ -13,6 +12,7 @@ import com.github.klee0kai.thekey.app.TargetDI
 import com.github.klee0kai.thekey.app.di.dependencies.AppComponentProviders
 import com.github.klee0kai.thekey.app.di.identifier.NoteIdentifier
 import com.github.klee0kai.thekey.app.di.identifier.StorageIdentifier
+import com.github.klee0kai.thekey.app.di.modules.AndroidHelpersModule
 import com.github.klee0kai.thekey.app.di.modules.CoroutineModule
 import com.github.klee0kai.thekey.app.di.modules.DBModule
 import com.github.klee0kai.thekey.app.di.modules.EngineModule
@@ -20,12 +20,9 @@ import com.github.klee0kai.thekey.app.di.modules.HelpersModule
 import com.github.klee0kai.thekey.app.di.modules.InteractorsModule
 import com.github.klee0kai.thekey.app.di.modules.PresentersModule
 import com.github.klee0kai.thekey.app.di.modules.RepositoriesModule
-import com.github.klee0kai.thekey.app.di.modules.AndroidHelpersModule
 import com.github.klee0kai.thekey.app.di.modules.ThemeModule
 import com.github.klee0kai.thekey.app.di.wrap.AppWrappersStone
 import com.github.klee0kai.thekey.app.model.AppConfig
-import com.github.klee0kai.thekey.app.ui.navigation.Destination
-import dev.olshevski.navigation.reimagined.NavController
 
 val DI: AppComponent = Stone.createComponent(AppComponent::class.java).apply {
     snackbarHostState(SnackbarHostState())
@@ -76,19 +73,10 @@ interface AppComponent : AppComponentProviders {
     @BindInstance(cache = BindInstance.CacheType.Weak)
     fun activity(app: Activity? = null): Activity?
 
-    @BindInstance
-    fun navigator(navController: NavController<Destination>? = null): NavController<Destination>
-
-    @BindInstance
-    fun backDispatcher(backDispatcher: OnBackPressedDispatcher? = null): OnBackPressedDispatcher
-
     @BindInstance(cache = BindInstance.CacheType.Strong)
     fun snackbarHostState(snackbarHostState: SnackbarHostState? = null): SnackbarHostState
-
 
     @BindInstance(cache = BindInstance.CacheType.Strong)
     fun config(snackbarHostState: AppConfig? = null): AppConfig
 
 }
-
-
