+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/ui/login/LoginPresenter.kt
@@ -1,15 +1,12 @@
 package com.github.klee0kai.thekey.app.ui.login
 
-import com.github.klee0kai.thekey.app.R
 import com.github.klee0kai.thekey.app.di.DI
 import com.github.klee0kai.thekey.app.di.identifier.StorageIdentifier
 import com.github.klee0kai.thekey.app.model.ColoredStorage
 import com.github.klee0kai.thekey.app.ui.navigation.StorageDestination
 import com.github.klee0kai.thekey.app.ui.navigation.StoragesDestination
-import com.github.klee0kai.thekey.app.ui.navigation.navigateForResult
-import com.github.klee0kai.thekey.app.ui.navigation.snack
+import com.github.klee0kai.thekey.app.ui.navigation.navigate
 import com.github.klee0kai.thekey.app.utils.coroutine.asyncResult
-import dev.olshevski.navigation.reimagined.navigate
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.flow
@@ -19,7 +16,7 @@ class LoginPresenter {
 
     private val storagesRep = DI.foundStoragesRepositoryLazy()
     private val settingsRep = DI.settingsRepositoryLazy()
-    private val navigator = DI.navigator()
+    private val router = DI.router()
     private val scope = DI.mainThreadScope()
 
     fun currentStorageFlow() = flow<ColoredStorage> {
@@ -30,8 +27,8 @@ class LoginPresenter {
     }
 
     fun selectStorage() = scope.launch {
-        val selectedStorage = navigator
-            .navigateForResult<String>(StoragesDestination)
+        val selectedStorage = router
+            .navigate<String>(StoragesDestination)
             .firstOrNull()
 
         if (selectedStorage != null) {
@@ -43,13 +40,13 @@ class LoginPresenter {
 
     fun login(passw: String) = scope.asyncResult {
         if (passw.isBlank()) {
-            navigator.snack(R.string.passw_is_null)
+//            router.snack(R.string.passw_is_null)
             return@asyncResult
         }
         val storage = currentStorageFlow().first()
         val engine = DI.cryptStorageEngineLazy(StorageIdentifier(storage.path))
         engine().login(passw)
-        navigator.navigate(StorageDestination(path = storage.path))
+        router.navigate(StorageDestination(path = storage.path))
     }
 
 }
