+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/ui/note/NotePresenter.kt
@@ -3,7 +3,6 @@ package com.github.klee0kai.thekey.app.ui.note
 import com.github.klee0kai.thekey.app.di.DI
 import com.github.klee0kai.thekey.app.di.identifier.StorageIdentifier
 import com.github.klee0kai.thekey.app.engine.model.DecryptedNote
-import com.github.klee0kai.thekey.app.ui.navigation.back
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
@@ -13,7 +12,7 @@ class NotePresenter(
     val notePtr: Long = 0,
 ) {
     private val scope = DI.mainThreadScope()
-    private val navigator = DI.navigator()
+    private val navigator = DI.router()
     private val engine = DI.cryptStorageEngineLazy(StorageIdentifier(path = storagePath))
 
     fun note(): Deferred<DecryptedNote> = scope.async {
