+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/ui/note/NoteScreen.kt
@@ -29,16 +29,15 @@ import com.github.klee0kai.thekey.app.di.identifier.NoteIdentifier
 import com.github.klee0kai.thekey.app.engine.model.DecryptedNote
 import com.github.klee0kai.thekey.app.ui.designkit.components.AppBarConst
 import com.github.klee0kai.thekey.app.ui.designkit.components.AppBarStates
-import com.github.klee0kai.thekey.app.ui.navigation.LocalNav
+import com.github.klee0kai.thekey.app.ui.navigation.LocalRouter
 import com.github.klee0kai.thekey.app.ui.navigation.NoteDestination
-import com.github.klee0kai.thekey.app.ui.navigation.back
 
 @Preview(showBackground = true)
 @Composable
 fun NoteScreen(
     args: NoteDestination = NoteDestination(),
 ) {
-    val navigator = LocalNav.current
+    val navigator = LocalRouter.current
     val presenter = remember { DI.notePresenter(NoteIdentifier(args.path, args.notePtr)) }
     val isEditNote = args.notePtr != 0L
     var isSkeleton by remember { mutableStateOf(isEditNote) }
