+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/ui/storage/StoragePresenter.kt
@@ -3,8 +3,6 @@ package com.github.klee0kai.thekey.app.ui.storage
 import com.github.klee0kai.thekey.app.di.DI
 import com.github.klee0kai.thekey.app.di.identifier.StorageIdentifier
 import com.github.klee0kai.thekey.app.engine.model.DecryptedNote
-import com.github.klee0kai.thekey.app.ui.navigation.StorageDestination
-import com.github.klee0kai.thekey.app.ui.navigation.awaitScreenEvent
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.flow
@@ -15,14 +13,14 @@ class StoragePresenter(
     val storagePath: String = "",
 ) {
     private val engine = DI.cryptStorageEngineLazy(StorageIdentifier(storagePath))
-    private val navigator = DI.navigator()
+    private val navigator = DI.router()
     private val scope = DI.mainThreadScope()
     private val updateTicks = MutableSharedFlow<Unit>()
 
     init {
         scope.launch {
-            navigator.awaitScreenEvent(StorageDestination(storagePath))
-            doLogout()
+//            navigator.awaitScreenEvent(StorageDestination(storagePath))
+//            doLogout()
         }
     }
