+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/ui/storage/genpassw/GeneratePasswordContent.kt
@@ -31,7 +31,7 @@ import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.constraintlayout.compose.Dimension
 import com.github.klee0kai.thekey.app.R
 import com.github.klee0kai.thekey.app.di.DI
-import com.github.klee0kai.thekey.app.ui.navigation.LocalNav
+import com.github.klee0kai.thekey.app.ui.navigation.LocalRouter
 
 @Preview
 @Composable
@@ -40,7 +40,7 @@ fun GeneratePasswordContent(
 ) {
     val scope = rememberCoroutineScope()
     val presenter = remember { DI.loginPresenter() }
-    val navigator = LocalNav.current
+    val navigator = LocalRouter.current
     val sliderValues = (4..12)
     var lenSliderPosition by remember { mutableIntStateOf(sliderValues.first) }
     var symbolsChecked by remember { mutableStateOf(false) }
