+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/ui/storages/components/StoragesListContent.kt
@@ -17,8 +17,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.github.klee0kai.thekey.app.R
 import com.github.klee0kai.thekey.app.di.DI
-import com.github.klee0kai.thekey.app.ui.navigation.LocalNav
-import com.github.klee0kai.thekey.app.ui.navigation.backWithResult
+import com.github.klee0kai.thekey.app.ui.navigation.LocalRouter
 import com.github.klee0kai.thekey.app.utils.views.animateAlphaAsState
 
 
@@ -28,7 +27,7 @@ fun StoragesListContent(
     modifier: Modifier = Modifier,
     showStoragesTitle: Boolean = true,
 ) {
-    val navigator = LocalNav.current
+    val navigator = LocalRouter.current
     val scope = rememberCoroutineScope()
     val presenter = remember { DI.storagesPresenter() }
     val storages = presenter.storages()
